bug fix and added new feature to IOUtil.php

+++ api/IOUtil.php
@@ -4,14 +4,25 @@ class IOUtil
 {
     public static $path = "/var/www/module2res/";
 
-    public static function saveFile($user, $file)
+    public static function saveFile($user, $file, $path="")
     {
-        $userPath = self::$path . $user . '/';
+        $userPath = self::$path . $user . '/' .$path;
         if (!file_exists($userPath)) {
             mkdir($userPath);
         }
         $savePath = $userPath . $file["name"];
         return move_uploaded_file($file["tmp_name"], $savePath);
+    }
+
+    public static function mkdir($user, $path, $currentPath="")
+    {
+        $userPath = self::$path . $user . '/' .$currentPath.'/'.$path.'/';
+        if (!file_exists($userPath)) {
+            mkdir($userPath);
+            return true;
+        } else {
+            return false;
+        }
     }
 
     public static function removeFile($user, $path)
@@ -22,8 +33,22 @@ class IOUtil
             echo "$filepath not exist";
             return;
         } else {
-            return unlink($filepath);
+            if (is_dir($filepath)) {
+                self::delTree($filepath);
+            } else {
+                return unlink($filepath);
+            }
         }
+    }
+
+    # credit https://www.php.net/manual/zh/function.rmdir.php#110489
+    public static function delTree($dir)
+    {
+        $files = array_diff(scandir($dir), array('.', '..'));
+        foreach ($files as $file) {
+            (is_dir("$dir/$file")) ? self::delTree("$dir/$file") : unlink("$dir/$file");
+        }
+        return rmdir($dir);
     }
 
     public static function listUserFiles($user, $innerPath = '')
@@ -33,7 +58,7 @@ class IOUtil
         if (!file_exists($userPath)) {
             mkdir($userPath);
         }
-        
+        $innerPath = rawurldecode($innerPath);
         $dirPath = self::formPath($user, $innerPath);
         $files = scandir($dirPath);
         return $files;
@@ -82,5 +107,22 @@ class IOUtil
     public static function formPath($user, $path)
     {
         return self::$path . $user . '/' . $path;
+    }
+
+    public static function unzip($user, $path)
+    {
+        $path = rawurldecode($path);
+        $filepath = self::formPath($user, $path);
+        $destpath = substr($filepath, 0, strrpos($filepath, '.')) . '/';
+        $zip = new ZipArchive;
+        $res = $zip->open($filepath);
+        if ($res === true) {
+            $zip->extractTo($destpath);
+            $zip->close();
+        }
+    }
+
+    public static function zip($user, $path){
+
     }
 }
